requirements.txt
Streamlitpymupdfopenaipython-dotenv
python-docx